+++ src/Contracts/Application/Ports/Driving/InboundEventDispatcher.php
@@ -10,11 +10,11 @@
 
 declare(strict_types=1);
 
-namespace CloudCreativity\Modules\Contracts\Application\Ports\Driving\InboundEvents;
+namespace CloudCreativity\Modules\Contracts\Application\Ports\Driving;
 
 use CloudCreativity\Modules\Contracts\Application\Messages\IntegrationEvent;
 
-interface EventDispatcher
+interface InboundEventDispatcher
 {
     /**
      * Dispatch an inbound integration event.